feat(mobile): Implement global image preview with advanced gestures
Refactor image preview functionality into a centralized provider:
- Create PreviewImageProvider for managing image preview state
- Add advanced gesture handling for pinch, pan, and tap interactions
- Improve image loading with transition and recycling key
- Enhance user experience with smooth animations and interactions
